+++ trunk/src/main/java/com/beimin/eveapi/eve/conquerablestationlist/StationListResponse.java
@@ -6,13 +6,13 @@ import java.util.Map;
 import com.beimin.eveapi.ApiResponse;
 
 public class StationListResponse extends ApiResponse {
-	private final Map<Integer, ApiStation> stationss = new HashMap<Integer, ApiStation>();
+	private final Map<Long, ApiStation> stationss = new HashMap<Long, ApiStation>();
 
 	public void addStation(ApiStation station) {
 		stationss.put(station.getStationID(), station);
 	}
 
-	public Map<Integer, ApiStation> getStations() {
+	public Map<Long, ApiStation> getStations() {
 		return stationss;
 	}
 }
